[Display] Improved display mode detection, FPS limit detection and many more display related functions. All backends benefit since the changes are SDL based.

+++ runtime/Core/Timer.h
@@ -41,8 +41,8 @@ namespace Spartan
         static void Tick();
 
         // FPS Limit
-        static void SetFpsLimit(double fps);
-        static double GetFpsLimit();
+        static void SetFpsLimit(float fps);
+        static float GetFpsLimit();
         static FpsLimitType GetFpsLimitType();
 
         // Times
